requirements.txt
adaptivecardbuilder==0.0.9
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backoff==2.1.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coloredlogs==15.0.1
Flask==2.1.2
frozenlist==1.3.0
future==0.18.2
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
urllib3==1.26.10
webex-bot==0.3.3
webexteamsbot==0.1.4.2
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.1
